route3_python_client/pyproject.toml
[tool.poetry]
name = "route3_client"
version = "0.1.0"
description = "api client for route3"
authors = ["Nico Mandery <nico@nmandery.net>"]

[tool.poetry.dependencies]
python = "^3.8"
grpcio = "^1.38.0"
Shapely = "^1.7.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
grpcio-tools = "^1.38.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
